Wrap Navi withRouter to give it access to history
