+++ frontend/components/header_and_footer/navi.jsx
@@ -1,124 +1,155 @@
-import React, {useState, useEffect, useRef} from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 
-function Navi(props) {
-
-    const [background, setBackground] = useState(false);
-    const [searchBarOpen, setSearchBarOpen] = useState(false);
-
-    const searchMovie = useRef();
-    const searchInput = useRef();
-
-    if (props.homepage) {
-        useEffect(() => {
-            window.addEventListener("scroll", () => {
-                if (window.scrollY > 50) {
-                    setBackground(true);
-                } else {
-                    setBackground(false);
-                }
-            });
-            return () => {
-                window.removeEventListener("scroll");
-            };
-        }, []);
+class Navi extends React.Component {
+    constructor(props) {
+        super(props)
+        this.state = {
+            background: false,
+            searchBarOpen: false,
+            queryString: "", 
+        };
+        this.searchInput = React.createRef();
+        this.handleClickClose = this.handleClickClose.bind(this);
+        this.handleClickOpen = this.handleClickOpen.bind(this);
+        this.handleScroll = this.handleScroll.bind(this);
+        this.search = this.search.bind(this);
+        this.searchDebounce = this.searchDebounce.bind(this);
+        this.endSession = this.endSession.bind(this);
+        this.updateValue = this.updateValue.bind(this);
+        this.test = this.test.bind(this);
     }
 
-    // useEventListener('mousedown', handleClickClose);
+    handleScroll() {
+        if (window.scrollY > 50) {
+            this.setState({ background: true });
+        } else {
+            this.setState({ background: false });
+        }
+    }
 
-    // function handleClickClose(e) {
-    //     if (searchMovie && !searchMovie.current.contains(e.target)) {
-    //     setSearchBarOpen(false);
-    //     searchInput.value = '';
-    //     }
-    // }
+    componentDidMount() {
+        if (this.props.homepage) {
+            window.addEventListener("scroll", this.handleScroll);
+        }
+    }
 
-    function handleClickOpen(e) {
-        if (searchBarOpen === false) {
-            setSearchBarOpen(true);
-            searchInput.current.focus();
+    componentWillUnmount() {
+        if (this.props.homepage) {
+            window.removeEventListener("scroll", this.handleScroll);
+        }
+    }
+
+    updateValue(field) {
+        return e => this.setState({
+            [field]: e.currentTarget.value
+        });
+    }
+
+    handleClickClose(e) {
+        this.setState({ queryString: "" })
+        this.setState({ searchBarOpen: false });
+    }
+
+    handleClickOpen(e) {
+        if (this.state.searchBarOpen === false) {
+            this.searchInput.current.focus();
+            this.setState({ searchBarOpen: true });
             e.target.value = '';
         }
     }
 
-    function formatSearch(input) {
-        const query = input.toLowerCase();
-        return `/search/${query}`;
-    }
-
-    function search(e) {
-        if (e.target.value === '') {
-            props.history.push('/browse');
+    search() {
+        if (this.state.queryString === '') {
+            this.props.history.push('/browse');
         } else {
-            props.history.push(formatSearch(e.target.value));
+            this.props.history.push(`search?q=${this.state.queryString}`);
         }
     }
 
-    function searchDebounce(e) {
-        setTimeout(() => search(e), 1000);
+    test(event) {
+        // event.persist();
+        console.log(event);
     }
 
-    function endSession() {
-        props.logout();
+    searchDebounce(e) {
+        // e.persist();
+        if (timer) {
+            clearTimeout(timer);
+        }
+        let timer = null;
+        timer = setTimeout(() => this.search(), 3000);
     }
 
-    if (props.loggedIn) {
-        return (
-            <header className={`pageHeader loggedIn ${background ? "active" : ""}`}>
-                <div className="firstSection">
-                    <Link to="/browse"><img src={window.logoURL} alt="FauxFlix Logo" className="logo"/></Link>
-                    <div className="subNavi">
-                        <Link to="/browse" className="headerLink">Home </Link>
-                        <Link to="/list" className="headerLink">My List </Link>
-                    </div>
-                </div>
-                <div className="secondSection">
+    endSession() {
+        this.props.logout();
+    }
 
-                    <div className="searchContainer">
-                        <div className="search" ref={searchMovie}>
-                            <input
-                                id="searchEle"
-                                onChange={searchDebounce}
-                                className={searchBarOpen ? 'openedInput' : 'closedInput'}
-                                type="text"
-                                placeholder="Movie Info"
-                                autoFocus
-                                ref={searchInput}>
-                            </input>
-                        <i onClick={handleClickOpen}
-                            className={searchBarOpen ? 'fas fa-search active sub-nav-logo' : 'fas fa-search sub-nav-logo'}>
-                        </i>
+    render() {
+        if (this.props.loggedIn) {
+            return (
+                <header className={`pageHeader loggedIn ${this.state.background ? "active" : ""}`}>
+                    <div className="firstSection">
+                        <Link to="/browse"><img src={window.logoURL} alt="FauxFlix Logo" className="logo"/></Link>
+                        <div className="subNavi">
+                            <Link to="/browse" className="headerLink">Home </Link>
+                            <Link to="/list" className="headerLink">My List </Link>
                         </div>
                     </div>
-
-                    <img className="profilePic" src={window.profile_pic} alt="Profile Pic" />
-                    <i className="fas fa-caret-down" id="downCaret"></i>
-                    <ul className="profileDropdown">
-                        <li className="dropdownItem">
-                            <a href="https://christophercfleming.github.io/" target="_blank">Portfolio</a>
-                        </li>
-                        <li className="dropdownItem">
-                            <a id="linkedin" href="https://www.linkedin.com/in/christophercfleming/" target="_blank">LinkedIn</a>
-                        </li>
-                        <li className="dropdownItem">
-                            <a id="github" href="https://github.com/ChristopherCFleming" target="_blank">Github</a>
-                        </li>
-                        <li className="dropdownItem secondToLast">
-                            <a href="https://angel.co/u/christopher-c-fleming" target="_blank">AngelList</a>
-                        </li>
-                        <li className="dropdownItem Last" onClick={endSession}>Logout</li>
-                    </ul>
-                </div>
-            </header>
-        )
-
-    } else {
-        return (
-            <header className="pageHeader">
-                <Link to="/"><img src={window.logoURL} alt="FauxFlix Logo" className="logo"/></Link>
-                <Link  id="login" className="btn" to="/login">Sign In</Link>
-            </header>
-        )
+                    <div className="secondSection">
+    
+                        <div className="searchContainer">
+                            <div className="search" ref={this.searchMovie}>
+                                <form onSubmit={this.search}>
+                                    <input
+                                        id="searchEle"
+                                        onChange={this.updateValue("queryString")}
+                                        className={this.state.searchBarOpen ? 'openedInput' : ''}
+                                        type="text"
+                                        placeholder="Title, Description"
+                                        autoFocus
+                                        value={this.state.queryString}
+                                        onInput={this.searchDebounce}
+                                        ref={this.searchInput}>
+                                        
+                                    </input>
+                                    <i onClick={this.handleClickOpen} className={this.state.searchBarOpen ? 'fas fa-search openedIcon' : 'fas fa-search'}/>
+                                    <i onClick={this.handleClickClose} id="exitIcon" className={this.state.searchBarOpen ? 'far fa-times-circle' : ''}></i>
+                                </form>
+                            </div>
+                        </div>
+                        <div className="profileSection">
+                            <img className="profilePic" src={window.profile_pic} alt="Profile Pic" />
+                            <i className="fas fa-caret-down" id="downCaret"></i>
+                            <ul className="profileDropdown">
+                                <li className="dropdownItem">
+                                    <a href="https://christophercfleming.github.io/" target="_blank">Portfolio</a>
+                                </li>
+                                <li className="dropdownItem">
+                                    <a id="linkedin" href="https://www.linkedin.com/in/christophercfleming/" target="_blank">LinkedIn</a>
+                                </li>
+                                <li className="dropdownItem">
+                                    <a id="github" href="https://github.com/ChristopherCFleming" target="_blank">Github</a>
+                                </li>
+                                <li className="dropdownItem secondToLast">
+                                    <a href="https://angel.co/u/christopher-c-fleming" target="_blank">AngelList</a>
+                                </li>
+                                <li className="dropdownItem Last" onClick={this.endSession}>Logout</li>
+                            </ul>
+                        </div>
+    
+                    </div>
+                </header>
+            )
+    
+        } else {
+            return (
+                <header className="pageHeader">
+                    <Link to="/"><img src={window.logoURL} alt="FauxFlix Logo" className="logo"/></Link>
+                    <Link  id="login" className="btn" to="/login">Sign In</Link>
+                </header>
+            )
+        }
     }
 }
