+++ frontend/components/search/search_container.js
@@ -1,4 +1,5 @@
 import { connect } from 'react-redux';
+import { withRouter } from 'react-router';
 import React from 'react';
 import Search from './search';
 import { fetchLists } from '../../actions/list_actions';
@@ -7,6 +8,9 @@ import { allVideos } from '../../actions/video_actions';
 const mapStateToProps = (state, ownProps) => {
     return {
         videosArray: Object.values(state.entities.videos),
+        query: ownProps.location.search.split("q=")[1],
+
+        
     };
 };
 
@@ -16,4 +20,4 @@ const mapDispatchToProps = dispatch => {
     };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(Search);
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(Search));
